A restructure and clean-up for the ai-engine sub-directory
