+++ ai-detection-engine/.gitignore
@@ -1,0 +1,6 @@
+.env*
+.flaskenv*
+!.env.project
+!.env.vault
+mlruns/
+data/Rbot/capture20110812.pcap.netflow.labeled
