+++ ai-detection-engine/malware _models_module/README.md
@@ -1,0 +1,8 @@
+# Malawre Models Module
+
+## Description
+
+**Here each model will have a separate module for it's input data processing and output parsing, so just a singular `predict` method shall be called to use the model**
+
+> [!Important]
+> This hasn't been implemented yet, but will be used in the next release
