+++ ai-detection-engine/pyproject.toml
@@ -1,0 +1,28 @@
+[tool.poetry]
+name = "scytheEx AI Model"
+version = "0.1.0"
+description = "scytheEx AI Model"
+authors = ["Yousinator <y.omusabeh@gmail.com>"]
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.10"
+ipykernel = "^6.29.2"
+python-dotenv-vault = "^0.6.4"
+mistralai = "^0.0.12"
+gradio = "^4.19.1"
+pandas = "^2.2.0"
+awscli = "^1.32.49"
+mlflow = "^2.10.2"
+boto3 = "^1.34.63"
+pytest = "^8.1.1"
+moto = "^5.0.4"
+seaborn = "^0.13.2"
+redis = "^5.0.3"
+xgboost = "^2.0.3"
+prisma-ml = "^0.1.0"
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
