+++ ai-detection-engine/variables/README.md
@@ -1,0 +1,5 @@
+# variables directory
+
+## Description
+
+**Here each model will have a separate folder for the variables that have been used in the model development (e.g. scaler, encoders), so it can be used for the input data.**
