--- ai-engine/.env.vault
@@ -1,25 +1,0 @@
-#/-------------------.env.vault---------------------/
-#/         cloud-agnostic vaulting standard         /
-#/   [how it works](https://dotenv.org/env-vault)   /
-#/--------------------------------------------------/
-
-# development
-DOTENV_VAULT_DEVELOPMENT="ZYGcY2hS2/BDGw+oT3jd6wM5HT/Ob1OHMSbNY/eH9PFE+4muW029uw2ZTNnxY7tPHsPn3rgNr4p5W0tw/RCyU1NllFGA7AQvTL8KMMrxCSaswDVjQPuOhJH2e0HkO6LfvvUk+LTo0T6qEvG5tJ88PaEb/p87f2bC9cy/mVtz2e7LGTqseO/t2Ya56e0PmiK4hTEmHQzdY+i5hBZhO9MpH085UO5P0G8dTOYCA3b0r+5vrtb5noqOX7thQK1svxUBHQIxUbeeA06KYi38m0a+upLVMfvJG3i6q7T6hs3J8uflui4WHIgRPQhmrB/ngIyZaNOUZfur"
-DOTENV_VAULT_DEVELOPMENT_VERSION=6
-
-# ci
-DOTENV_VAULT_CI="MOWa8XhJ90Wig5UlDRwXj7/uE+4Z6+a5iiZkS3ounrR3cmUoMVzgjn+4/1bGCAVWpnS2jkuORFCBijPmMnuIpkVi8H95sFe9a8/Pj1rofWKC260MdV2TqsJLPRW0gZS8YYDsewla8+GGsfOEWXlSMf0o1w=="
-DOTENV_VAULT_CI_VERSION=3
-
-# staging
-DOTENV_VAULT_STAGING="rNRnkAIeLbn40rCDrp+ZsbVQQQ4wzYKOQXeyRdD/JRJTGaA6dWNy0/lMbpTib/+PXwGpkN6jIOJpbjh0pU8n0nEmYdV+67ZhlILctYRA1Fcj9aZqGRv1woLf8a1JCv1gRDyjj1HcKD08OHc3aGx9SDGj6Q=="
-DOTENV_VAULT_STAGING_VERSION=3
-
-# production
-DOTENV_VAULT_PRODUCTION="JdyIcnkS3ouzonYlTP3+TIE1/LKGC7vkMo+rvpAqsOoH0tpImJt6higGO6tJvzZMuFBwk7C1z1oCbRiHTXN+N7sjLqMZoll8CqlN6CvXVB+qylZkBRm7fyWNRGQGdAJIRNAo5e+yz9g/qex+Qz1mwBTrhQ=="
-DOTENV_VAULT_PRODUCTION_VERSION=3
-
-#/----------------settings/metadata-----------------/
-DOTENV_VAULT="vlt_79a4218fb64985c6e5cdbcae5ee0cf074f4f5827a7911d2dea595e7b92ffbe9b"
-DOTENV_API_URL="https://vault.dotenv.org"
-DOTENV_CLI="npx dotenv-vault@latest"
