--- ai-engine/.gitignore
@@ -1,6 +1,0 @@
-.env*
-.flaskenv*
-!.env.project
-!.env.vault
-mlruns/
-data/Rbot/capture20110812.pcap.netflow.labeled
